Effects progress, playback speed, and some utility classes

+++ src/main/java/org/emrick/project/SyncListener.java
@@ -3,5 +3,5 @@ package org.emrick.project;
 import java.util.ArrayList;
 
 public interface SyncListener {
-    void onSync(ArrayList<SyncTimeGUI.Pair> times, int startDelay);
+    void onSync(ArrayList<SyncTimeGUI.Pair> times, float startDelay);
 }
